Rename user_config directory to general_config directory.
