+++ indy_common/config.py
@@ -102,5 +102,3 @@ agentLoggingLevel = logging.INFO
 default logging level for node
 '''
 logLevel = logging.INFO
-
-USER_CONFIG_DIR = "none"
