+++ indy_node/general_config/windows_platform_config.py
